requirements.txt
simpletransformers==0.63.6
transformers==4.17.0
pandas
tqdm
matplotlib
wandb
pyyaml
